# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "Dokumentasi API SiMas",
        "contact": {
            "email": "example@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Dokumentasi API SiMas"
        }
    ],
    "paths": {
...
        "/warga": {
            "get": {
                "tags": [
                    "Warga"
                ],
                "summary": "Get all warga data",
                "operationId": "9b36486c1a9c59847bb85d49ae271cca",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved all warga data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success retrieve all warga data"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Warga"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Warga"
                ],
                "summary": "Store new warga",
                "operationId": "985009373c4c5a81dcba0eeb46748b2c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nik",
                                    "rt_id",
                                    "name",
                                    "address"
                                ],
                                "properties": {
                                    "nik": {
                                        "type": "string",
                                        "example": "3026300139170900"
                                    },
                                    "birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-01-01"
                                    },
                                    "rt_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Magdalen Gleichner V"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "1626 Jast Keys Apt. 210\\nMagnusside, ND 33533"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Warga stored successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success store warga"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Warga"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Your input is invalid"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "nik": [
                                                    "The nik must be at least 15 digits."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "SQLSTATE[23000]: Integrity constraint violation: ..."
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Warga": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nik": {
                        "type": "string",
                        "example": "3026300139170900"
                    },
                    "rt_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Magdalen Gleichner V"
                    },
                    "birth": {
                        "type": "string",
                        "format": "date",
                        "example": "2022-01-01"
                    },
                    "address": {
                        "type": "string",
                        "example": "1626 Jast Keys Apt. 210\\nMagnusside, ND 33533"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-03T20:22:05.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-03T20:22:05.000000Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}